This merge completes the following 1. Simplifies the implementation by merging batch_prover.rs & batch_verifier.rs functionality into prover.rs & verifier.rs 2. Improves the performance tests
