+++ marlin-succinct/protocol/src/lib.rs
@@ -1,4 +1,2 @@
 pub mod prover;
 pub mod verifier;
-pub mod batch_prover;
-pub mod batch_verifier;
